requirements.txt
python-kucoin
TA-Lib
numpy
pandas
websocket_client
pprint